refactored database code
