+++ mailer/email_formatter.py
@@ -4,9 +4,9 @@ class EmailTemplateRenderer:
     def __init__(self, template_dir: str) -> None:
         self.env = Environment(loader=FileSystemLoader(template_dir))
 
-    def render_template(self, template_name: str, recipient_email: str, **kwargs: dict) -> str:
+    def render_template(self, template_name: str, **kwargs: dict) -> str:
         template = self.env.get_template(template_name)
-        rendered_content = template.render(recipient_email=recipient_email, **kwargs)
+        rendered_content = template.render(**kwargs)
         return rendered_content
