+++ nifty.py
@@ -1,12 +1,12 @@
 import argparse
 import os
+
 from cloud.file_upload import FileUploaderClass
 from zipper.file_zipper import FileZipper
 from mailer.email_formatter import EmailTemplateRenderer
 from mailer.email_sender import EmailSender
-from database.database import DatabaseFactory
-
 from settings import MailerConfig, DatabaseConfig
+from database.database import DatabaseFactory, get_db_connection
 
 import logging
 import sys
@@ -24,41 +24,6 @@ logging.basicConfig(
 )
 
 logger = logging.getLogger(__name__)
-
-
-def get_db_connection():
-    
-    db_config = DatabaseConfig()
-
-    if db_config.DB_TYPE == "mysql":
-        mysql_db = DatabaseFactory.create_database(
-            db_type="mysql",
-            host=db_config.MYSQL_DB_HOST,
-            user=db_config.MYSQL_DB_USERNAME,
-            password=db_config.MYSQL_DB_PASSWORD,
-            database=db_config.MYSQL_DB_NAME)
-
-        with mysql_db as db:
-            db.create_table("transfers", db_config.MYSQL_COLS)
-
-        return mysql_db
-
-    elif db_config.DB_TYPE == "sqlite":
-        db_file=db_config.SQLITE_DB_FILENAME
-        if not os.path.exists(db_file):
-
-            sqlite_db = DatabaseFactory.create_database(
-                db_type="sqlite",
-                db_file=db_config.SQLITE_DB_FILENAME)
-
-            with sqlite_db as db:
-                db.create_table("transfers", db_config.SQLITE_COLS)
-
-        else:
-            sqlite_db = DatabaseFactory.create_database(
-                db_type="sqlite",
-                db_file=db_config.SQLITE_DB_FILENAME)
-        return sqlite_db
 
 
 if __name__ == '__main__':
@@ -99,22 +64,25 @@ if __name__ == '__main__':
     else:
         plural = "File"
 
-    mail_context = {
+    data_params = {
         "sender_name": MailerConfig.MAIL_HOST_SENDER_NAME,
         "file_basename": os.path.basename(prepped_file_path),
         "sender_address": MailerConfig.MAIL_HOST_SENDER_ADDRESS,
         "download_link": link,
-        "recipent_email": args.recipient,
+        "recipient_email": args.recipient,
+        "expiry_date_dt": expiry_date_dt,
         "expiry_date": expiry_date,
+        "file_size_bytes": os.path.getsize(prepped_file_path),
         "file_size_mb": round(os.path.getsize(prepped_file_path) / 1024 / 1024, 2),
         "files_list": files_list,
         "file_count": file_count,
-        "plural": plural
+        "plural": plural,
+        "local_datetime": str(datetime.now())[:19]
     }
 
     logger.info(f"Rendering email template {args.template}")
     template_renderer = EmailTemplateRenderer(template_dir='mail_templates')
-    email_content = template_renderer.render_template(args.template, args.recipient, **mail_context)
+    email_content = template_renderer.render_template(args.template, **data_params)
 
     # Email send
     logger.info(f"Sending email to {args.recipient}")
@@ -123,23 +91,9 @@ if __name__ == '__main__':
 
     # Database Insert
     logger.info(f"Logging to database")
-        
-    connection = get_db_connection()
-    
+    connection = get_db_connection() 
     with connection as db:
-        db.insert_data(
-            "transfers",
-            {
-                "sender_name": MailerConfig.MAIL_HOST_SENDER_NAME,
-                "file_basename": os.path.basename(prepped_file_path),
-                "sender_address": MailerConfig.MAIL_HOST_SENDER_ADDRESS,
-                "download_link": link,
-                "recipient_email": args.recipient,
-                "expiry_date": expiry_date_dt,
-                "file_size_mb": round(os.path.getsize(prepped_file_path) / 1024 / 1024, 2),
-                "files_list": ", ".join(files_list)
-            }
-        )
+        db.insert_data("transfers", data_params)
 
     # Clean up
     if prepped_file_path.endswith('.zip'):
